+++ Changes.md
@@ -5,6 +5,7 @@ Features
 --------
 
 - FileList : Added node for listing matching files.
+- DeleteFiles : Added node for deleting files.
 
 Improvements
 ------------